+++ src/index.ts
@@ -1,5 +1,3 @@
-import './fetch';
-
 import {
   Request,
   createNetworkInterface,
