+++ test/tests.ts
@@ -53,4 +53,3 @@ import './customResolvers';
 import './isEqual';
 import './cloneDeep';
 import './assign';
-import './fetch';
